requirements.txt
langchain-core
langchain-openai
langchain-community
python-dotenv
SpeechRecognition
python-dateutil
openai
pyaudio
sqlite3